requirements.txt
opencv-python-headless
numpy
pytesseract
Pillow
pydicom